+++ src/QuestFrameworkMod.cs
@@ -35,7 +35,7 @@ namespace QuestFramework
         internal QuestManager QuestManager { get; private set; }
         internal QuestStateStore QuestStateStore { get; private set; }
         internal StatsManager StatsManager { get; private set; }
-        internal HookManager HookManager { get; private set; }
+        internal ConditionManager HookManager { get; private set; }
         internal QuestOfferManager QuestOfferManager { get; private set; }
         internal Loader ContentPackLoader { get; private set; }
         internal QuestController QuestController { get; private set; }
@@ -59,7 +59,7 @@ namespace QuestFramework
             this.QuestManager = new QuestManager(this.Monitor);
             this.QuestStateStore = new QuestStateStore(helper.Data, this.Monitor);
             this.StatsManager = new StatsManager(helper.Multiplayer, helper.Data);
-            this.HookManager = new HookManager(this.Monitor);
+            this.HookManager = new ConditionManager(this.Monitor);
             this.QuestOfferManager = new QuestOfferManager(this.HookManager, this.QuestManager);
             this.ContentPackLoader = new Loader(this.Monitor, this.QuestManager, this.QuestOfferManager);
             this.QuestController = new QuestController(this.QuestManager, this.QuestOfferManager, this.Monitor);